refactor: move rich text into a subfolder, to prepare for more components
