+++ src/rich-text/__tests__/style-object.test.ts
